+++ crates/astria-sequencer/src/metrics.rs
@@ -42,6 +42,7 @@ pub struct Metrics {
     extend_vote_failure_count: Counter,
     verify_vote_extension_failure_count: Counter,
     ibc_relay_failures: Histogram,
+    results_in_recently_executed_cache: Gauge,
 }
 
 impl Metrics {
@@ -166,6 +167,10 @@ impl Metrics {
 
     pub(crate) fn increment_verify_vote_extension_failure_count(&self) {
         self.verify_vote_extension_failure_count.increment(1);
+    }
+
+    pub(crate) fn set_results_in_recently_executed_cache(&self, count: usize) {
+        self.results_in_recently_executed_cache.set(count);
     }
 
     pub(crate) fn record_ibc_relay_failures(&self, count: usize) {
@@ -377,6 +382,13 @@ impl telemetry::Metrics for Metrics {
             )?
             .register()?;
 
+        let results_in_recently_executed_cache = builder
+            .new_gauge_factory(
+                RESULTS_IN_RECENTLY_EXECUTED_CACHE,
+                "The number of results in the recently executed results cache",
+            )?
+            .register()?;
+
         let ibc_relay_failures = builder
             .new_histogram_factory(
                 IBC_RELAY_FAILURES,
@@ -413,6 +425,7 @@ impl telemetry::Metrics for Metrics {
             extend_vote_duration_seconds,
             extend_vote_failure_count,
             verify_vote_extension_failure_count,
+            results_in_recently_executed_cache,
             ibc_relay_failures,
         })
     }
@@ -445,6 +458,7 @@ metric_names!(const METRICS_NAMES:
     EXTEND_VOTE_DURATION_SECONDS,
     EXTEND_VOTE_FAILURE_COUNT,
     VERIFY_VOTE_EXTENSION_FAILURE_COUNT,
+    RESULTS_IN_RECENTLY_EXECUTED_CACHE,
     IBC_RELAY_FAILURES,
 );
 
@@ -519,6 +533,10 @@ mod tests {
         assert_const(
             VERIFY_VOTE_EXTENSION_FAILURE_COUNT,
             "verify_vote_extension_failure_count",
+        );
+        assert_const(
+            RESULTS_IN_RECENTLY_EXECUTED_CACHE,
+            "results_in_recently_executed_cache",
         );
         assert_const(IBC_RELAY_FAILURES, "ibc_relay_failures");
     }